requirements.txt
yfinance
pyyaml
dotenv
